+++ runtime/src/iree/builtins/ukernel/api.h
@@ -7,7 +7,6 @@
 #ifndef IREE_BUILTINS_UKERNEL_API_H_
 #define IREE_BUILTINS_UKERNEL_API_H_
 
-#include "iree/builtins/ukernel/elementwise.h"
 #include "iree/builtins/ukernel/mmt4d.h"
 #include "iree/builtins/ukernel/pack.h"
 #include "iree/builtins/ukernel/query_tile_sizes.h"
