+++ runtime/src/iree/modules/vmvx/BUILD.bazel
@@ -15,6 +15,8 @@ package(
 iree_runtime_cc_library(
     name = "vmvx",
     srcs = [
+        "elementwise.c",
+        "elementwise.h",
         "module.c",
     ],
     hdrs = [
