+++ runtime/src/iree/modules/vmvx/CMakeLists.txt
@@ -17,7 +17,8 @@ iree_cc_library(
   TEXTUAL_HDRS
     "exports.inl"
   SRCS
-    "module.c"
+  "elementwise.c"
+  "module.c"
   DEFINES
     "IREE_HAVE_VMVX_MODULE"
   DEPS
